+++ front/src/views/CategoryView.vue
@@ -125,6 +125,7 @@ watch(
     :category="category"
     :thumbnail-orientation="parameters.thumbnailOrientation.value"
     :thumbnail-image-fit="parameters.thumbnailImageFit.value"
+    :thumbnail-size-multiplier="parameters.thumbnailSizeMultiplier.value"
     @select-item="emit('select-item', $event)"
     @select-category="emit('select-category', $event)"
     @update:background-media-items="emit('update:background-media-items', $event)"